refactor: 입력받는 데이터 전달용PostWriteForm dto 사용하도록 수정
글쓰기 시 입력하는 데이터 전달용 dto인 PostWriteForm을 생성했습니다.
PostWriterForm은 BoardPost 객체로 만들어 반환해주는 toBoardPost()를 가지고 있습니다.
Controller에서 데이터를 전달 받아 PostWriteForm을 생성하고 BoardRepository로 전달합니다.
BoardRepository는 PostWriteForm을 BoardPost로 변환하여 저장하도록 구현했습니다.
